extract common tools from fast-filters
